Rename Report.status to Report.check_status to avoid name shadowing (#128)
In the current code, `Report.status` refers to two things:
- A class method to check the status of a report
- The status attribute of a `Report` object

The problem is that instances of the class also have class methods,
which introduces confusion, as accessing `status` on an instance will
refer to:
- the attribute if it has been set by the API response
- the class method otherwise
When accessing `status` on an instance, VS Code's static code analysis
identifies it as the class method.

This renames the class method to avoid confusion.